Change suitors to be built as array w/ varied size
Previous construction was throwing results generation for k-modes paper
off for certain runs. This was because taking just the k most similar
points to each centroid meant that for dense datasets (like the zoo
animal set) there would be less than k unique points chosen as suitors
in the matching game. In this way, not all reviewers (centroids) would
be matched to a point in the data and the k-modes algorithm would fail
due to the presence of empty clusters.

To remedy this, at the expense of increased computational time, instead
of k points being chosen, we take min(N, k^2). This (so far from
practical tests) stops the error occurring with little impact on the
other performance metrics of the algorithm.

It should also be noted that once these k * min(N, k^2) potential
suitors are found, we apply np.unique on them so we avoid empty initial
clusters due to identical centroids being chosen.

This feels like a bit of a hacky fix at the moment since min(N, k^2) was
chosen to be the size of the suitor set arbitrarily/to fix the current
problem. This value could well be far from optimal, and is very much
subject to change.