Make matching inits in line with matching algrithm
Found that the capacitated Gale-Shapley algorithm was in fact inaccurate
(this has now been fixed) so things have been updated accordingly here.
This has led to overall improvements to this initialisation method
across several test datasets in terms of final costs, run-time, and some
clustering metrics.

For the sake of readability and functionality, reviewers and suitors are
kept as tuples in respective lists rather than numpy arrays to be
converted back and forth between arrays and tuples.