Add class to deal with duplicate dict keys
Sometimes random sampling of category values in Huang's method, and more
importantly ours, results in duplicate virtual modes being selected. As
Python does not support duplicate keys in dictionaries, a simple class
`DataPoint` is made with one attribute `attrs` which is the value of the
underlying point's attributes. This is a bit ugly and will likely need
brushing up. However, there are no limitations on the seeds (or number
of initialisations) we can use matching initialisations for.